% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_analyse.R
\name{quick_analyse}
\alias{quick_analyse}
\title{Quickly analyse a data frame using h2o.ai}
\usage{
quick_analyse(data, response_variable = "y", analysis_mode = c("light",
  "full"))
}
\arguments{
\item{data}{input data frame to be analysed}

\item{response_variable}{name (string) of column in the data frame that
contains the response / dependent variable. If missing, defaults to "y".}

\item{analysis_mode}{one of either "light" (default) or "full". This determines
the options that will be passed to the internal models. Light mode aims to
reduce run time at the expense of accuracy, full mode is more heavyweight.
Your mileage may vary.}
}
\value{
a list, each item of which is another list containing the h2o model
  objects, and the h2o.performance objects extracted from those objects,
  respectively.
}
\description{
Uses h2o.ai to quickly run a random forest, gradient boosting machine and
  deep learning model on an input data frame, and returns information on the
  performance of the models.
}
\examples{
data <- datasets::iris
quick_result <- quick_analyse(data = data, response_variable = "Species")
full_result <- quick_analyse(data = data, response_variable = "Species", analysis_mode = "full")
}
